+++ app/(chat)/page.tsx
@@ -1,12 +1,7 @@
 import { nanoid } from '@/lib/utils'
 import { Chat } from '@/components/chat'
 import { AI } from '@/lib/chat/actions'
-import { Session } from '@/lib/types'
 import { getMissingKeys } from '@/app/actions'
-
-export const metadata = {
-  title: '888 StockBot'
-}
 
 export default async function IndexPage() {
   const id = nanoid()